Create component folder and move files to match develop branch
